requirements_ak.txt
aioredis==1.3.1
amqp==2.6.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.2
Automat==20.2.0
beautifulsoup4==4.9.1
billiard==3.6.4.0
bs4==0.0.1
celery==4.4.5
certifi==2020.4.5.1
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
daphne==3.0.1
dateutils==0.6.8
Django==3.0.6
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-redis==4.12.1
django-redis-sessions==0.6.2
djangorestframework==3.12.2
enum34==1.1.10
fake-useragent==0.1.11
feedparser==6.0.8
future==0.18.2
googlefinance==0.7
gunicorn==20.0.4
h11==0.12.0
h2==4.0.0
hiredis==1.1.0
hpack==4.0.0
html5lib==1.0.1
httptools==0.1.1
Hypercorn==0.11.2
hyperframe==6.0.0
hyperlink==20.0.1
idna==2.9
importlib-metadata==1.6.1
incremental==17.5.0
iniconfig==1.1.1
ipynb-py-convert==0.4.5
kiteconnect==3.9.0
kombu==4.6.11
lxml==4.5.1
msgpack==1.0.2
multitasking==0.0.9
nest-asyncio==1.4.3
newsapi-python==0.2.6
nsepy==0.8
nsetools==1.0.11
numpy==1.18.5
packaging==20.8
pandas
pandas-datareader==0.8.1
parse==1.15.0
pluggy==0.13.1
priority==1.3.0
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyee==7.0.2
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
pytest==6.2.1
python-dateutil==2.8.1
pytz==2020.1
razorpay==1.2.0
redis==3.5.3
requests==2.23.0
requests-html==0.10.0
selenium==3.141.0
sentry-sdk==1.0.0
service-identity==18.1.0
sgmllib3k==1.0.0
simplejson==3.17.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.2
tqdm==4.46.1
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
urllib3==1.25.9
uvicorn==0.13.3
uvloop==0.15.1
vine==1.3.0
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
wsproto==1.0.0
yahoo-fin==0.8.9.1
yahoo-finance==1.4.0
yfinance==0.1.63
zipp==3.1.0
zope.interface==5.2.0
django-import-export==2.5.0
Pyrebase4==4.5.0